recording PR comment to refactor code
